refactor: 오늘 스케줄 컴포넌트 정리(DURI-194)
